[Refactor] 필요없는 메소드 삭제 및 메소드 수정
