+++ src/main/java/com/nhnacademy/bookpubshop/product/relationship/service/ProductSaleStateCodeService.java
@@ -1,6 +1,5 @@
 package com.nhnacademy.bookpubshop.product.relationship.service;
 
-import com.nhnacademy.bookpubshop.product.relationship.dto.CreateProductSaleStateCodeRequestDto;
 import com.nhnacademy.bookpubshop.product.relationship.dto.GetProductSaleStateCodeResponseDto;
 import java.util.List;
 
@@ -11,14 +10,6 @@ import java.util.List;
  * @since : 1.0
  */
 public interface ProductSaleStateCodeService {
-
-    /**
-     * 상품판매상태코드를 생성합니다.
-     *
-     * @param request 생성을 위한 Dto class.
-     * @return 생성된 객체를 반환합니다.
-     */
-    GetProductSaleStateCodeResponseDto createSaleCode(CreateProductSaleStateCodeRequestDto request);
 
     /**
      * 코드번호로 상품판매유형코드를 반환합니다.
